Change Vitest to Mocha for testing
