+++ package.json
@@ -7,16 +7,16 @@
   "type": "module",
   "scripts": {
     "example": "node ./tests/test.js",
-    "test": "vitest"
+    "test": "mocha --ui qunit --timeout 10000"
   },
   "dependencies": {
     "pyodide": "0.26.2"
   },
   "devDependencies": {
-    "vitest": "3.0.8",
     "eslint": "^8.0.0",
     "husky": "^8.0.0",
-    "lint-staged": "^14.0.0"
+    "lint-staged": "^14.0.0",
+    "mocha": "^10.2.0"
   },
   "husky": {
     "hooks": {
